Revert "Implement ZipkinService getTraceId functionality (#351)"
This reverts commit 7bccb333d01bdf3214ce13a1dd059754e6fa5ea1.
