+++ kaldb/src/main/java/com/slack/kaldb/zipkinApi/ZipkinService.java
@@ -1,0 +1,64 @@
+package com.slack.kaldb.zipkinApi;
+
+import com.linecorp.armeria.common.HttpResponse;
+import com.linecorp.armeria.common.HttpStatus;
+import com.linecorp.armeria.common.MediaType;
+import com.linecorp.armeria.server.annotation.*;
+import com.slack.kaldb.server.KaldbQueryServiceBase;
+import java.io.IOException;
+import java.util.*;
+
+@SuppressWarnings("OptionalUsedAsFieldOrParameterType")
+/**
+ * Zipkin compatible API service, for use in Grafana
+ *
+ * @see <a
+ *     href="https://github.com/grafana/grafana/blob/main/public/app/plugins/datasource/zipkin/datasource.ts">Grafana
+ *     Zipkin API</a> <a
+ *     href="https://github.com/openzipkin/zipkin-api/blob/master/zipkin.proto">Trace proto
+ *     compatible with Zipkin</a> <a href="https://zipkin.io/zipkin-api/#/">Trace API Swagger
+ *     Hub</a>
+ */
+public class ZipkinService {
+
+  private final KaldbQueryServiceBase searcher;
+
+  public ZipkinService(KaldbQueryServiceBase searcher) {
+    this.searcher = searcher;
+  }
+
+  @Get
+  @Path("/api/v2/services")
+  public HttpResponse getServices() throws IOException {
+    String output = "[]";
+    return HttpResponse.of(HttpStatus.OK, MediaType.JSON_UTF_8, output);
+  }
+
+  @Get("/api/v2/spans")
+  public HttpResponse getSpans(@Param("serviceName") Optional<String> serviceName)
+      throws IOException {
+    String output = "[]";
+    return HttpResponse.of(HttpStatus.OK, MediaType.JSON_UTF_8, output);
+  }
+
+  @Get("/api/v2/traces")
+  public HttpResponse getTraces(
+      @Param("serviceName") Optional<String> serviceName,
+      @Param("spanName") Optional<String> spanName,
+      @Param("annotationQuery") Optional<String> annotationQuery,
+      @Param("minDuration") Optional<Integer> minDuration,
+      @Param("maxDuration") Optional<Integer> maxDuration,
+      @Param("endTs") Long endTs,
+      @Param("lookback") Long lookback,
+      @Param("limit") @Default("10") Integer limit)
+      throws IOException {
+    String output = "[]";
+    return HttpResponse.of(HttpStatus.OK, MediaType.JSON_UTF_8, output);
+  }
+
+  @Get("/api/v2/trace/{traceId}")
+  public HttpResponse getTraceByTraceId(@Param("traceId") String traceId) throws IOException {
+    String output = "[]";
+    return HttpResponse.of(HttpStatus.OK, MediaType.JSON_UTF_8, output);
+  }
+}
