+++ kaldb/src/test/java/com/slack/kaldb/chunkManager/RecoveryChunkManagerTest.java
@@ -120,8 +120,7 @@ public class RecoveryChunkManagerTest {
             KaldbConfigs.NodeRole.RECOVERY,
             10000,
             "api_log",
-            9003,
-            100);
+            9003);
 
     chunkManager =
         RecoveryChunkManager.fromConfig(
