+++ kaldb/src/test/java/com/slack/kaldb/testlib/KaldbConfigUtil.java
@@ -3,7 +3,6 @@ package com.slack.kaldb.testlib;
 import com.slack.kaldb.proto.config.KaldbConfigs;
 
 public class KaldbConfigUtil {
-
   public static KaldbConfigs.KaldbConfig makeKaldbConfig(
       String bootstrapServers,
       int indexerPort,
@@ -17,8 +16,7 @@ public class KaldbConfigUtil {
       KaldbConfigs.NodeRole nodeRole,
       int maxOffsetDelay,
       String dataTransformerConfig,
-      int recoveryPort,
-      int maxMessagesPerChunk) {
+      int recoveryPort) {
     KaldbConfigs.KafkaConfig kafkaConfig =
         KaldbConfigs.KafkaConfig.newBuilder()
             .setKafkaTopic(kafkaTopic)
@@ -47,7 +45,7 @@ public class KaldbConfigUtil {
                     .setRequestTimeoutMs(5000)
                     .build())
             .setMaxBytesPerChunk(10L * 1024 * 1024 * 1024)
-            .setMaxMessagesPerChunk(maxMessagesPerChunk)
+            .setMaxMessagesPerChunk(100)
             .setLuceneConfig(
                 KaldbConfigs.LuceneConfig.newBuilder()
                     .setCommitDurationSecs(10)
